data should now be persistent (untested) i'd like to test it but the unit tests make it difficult as it involves the whole program quitting and restarting also added a unit test, cleaned the other one, changed the maximal size of events from UInt16.MaxSize to (normally) Int32.MaxSize (at the cost of 2B/event), switched dummy mode to true for back unit tests (i didn't notice it, i should have done it earlier) and changed two things a bit dangerous in bufferedblobio

+++ EvernestBack/WriteLocker.cs
@@ -34,8 +34,8 @@ namespace EvernestBack
                     {
                         var agent = new Agent(pendingEvent.Message, CurrentID, pendingEvent.CallbackSuccess,
                             pendingEvent.CallbackFailure);
-                        ulong wroteBytes = Write(agent);
-                        _indexer.NotifyNewEntry(CurrentID, wroteBytes);
+                        int wroteBytes = Write(agent);
+                        _indexer.NotifyNewEntry(CurrentID, (ulong) wroteBytes);
                         CurrentID++;
                         agent.Processed();
                     }
@@ -44,9 +44,9 @@ namespace EvernestBack
                 );
         }
 
-        private UInt16 Write(Agent prod)
+        private int Write(Agent prod)
         {
-            UInt16 size;
+            int size;
             var bytes = prod.Serialize(out size);
             _writeBuffer.Push(bytes, 0, size);
             return size;
